Lots of stuff for pause
I decided to go with the pause that just draws the background without most objects getting drawn. It will specifically draw snow, and anything with the wallsParent object as it's parent. I have created an object for the restart button and sprites for the main menu and continue button. Currently still does not draw anything besides the background for the pause menu, and the pause menu looks very low quality but I am uncertain on if I could even fix that since the rest of the game is that quality when in levels.